New folder structure and webpack config WIP.
