+++ public/js/lib/confirmation-modal.js
