+++ public/js/lib/metabox-autocomplete.js
